+++ modules/config/config.go
@@ -4,6 +4,7 @@ import (
 	"github.com/spf13/viper"
 )
 
+// initializeConfig initialized viper config using yaml config
 func (c *Config) initalizeConfig() error {
 	viper.SetConfigFile("config.yaml")
 	if err := viper.ReadInConfig(); err != nil {
