+++ modules/module.go
@@ -2,9 +2,11 @@ package modules
 
 import (
 	"github.com/minish144/rentateam-test/modules/config"
+	"github.com/minish144/rentateam-test/modules/logger"
 	"go.uber.org/fx"
 )
 
 var Module = fx.Options(
 	config.Module,
+	logger.Module,
 )
